Move Genesis tests to `ouroboros-consensus-diffusion:consensus-test`
Also move the two `Extras` modules into `Test.Consensus.*`.
