+++ ouroboros-consensus-diffusion/test/consensus-test/Test/Consensus/Genesis/Tests.hs
@@ -1,0 +1,9 @@
+module Test.Consensus.Genesis.Tests (tests) where
+
+import qualified Test.Consensus.Genesis.Tests.LongRangeAttack as LongRangeAttack
+import           Test.Tasty
+
+tests :: TestTree
+tests = testGroup "Genesis tests"
+    [ LongRangeAttack.tests
+    ]
