+++ ouroboros-consensus-diffusion/test/consensus-test/Test/Consensus/PeerSimulator/Run.hs
@@ -4,7 +4,7 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeFamilies        #-}
 
-module Test.Ouroboros.Consensus.PeerSimulator.Run (
+module Test.Consensus.PeerSimulator.Run (
     ChainSyncException (..)
   , runPointSchedule
   ) where
@@ -43,7 +43,6 @@ import           Ouroboros.Network.Channel (createConnectedChannels)
 import           Ouroboros.Network.ControlMessage (ControlMessage (..),
                      ControlMessageSTM)
 import           Ouroboros.Network.Driver.Limits
-import           Ouroboros.Network.Driver.Limits.Extras
 import           Ouroboros.Network.Driver.Simple (Role (Client, Server))
 import           Ouroboros.Network.Protocol.ChainSync.ClientPipelined
                      (ChainSyncClientPipelined, chainSyncClientPeerPipelined)
@@ -52,18 +51,18 @@ import           Ouroboros.Network.Protocol.ChainSync.PipelineDecision
                      (pipelineDecisionLowHighMark)
 import           Ouroboros.Network.Protocol.ChainSync.Server
                      (chainSyncServerPeer)
-import qualified Test.Ouroboros.Consensus.BlockTree as BT
+import qualified Test.Consensus.BlockTree as BT
+import           Test.Consensus.Genesis.Setup.GenChains (GenesisTest)
+import           Test.Consensus.Network.Driver.Limits.Extras
+import qualified Test.Consensus.PeerSimulator.BlockFetch as PeerSimulator.BlockFetch
+import           Test.Consensus.PeerSimulator.Config
+import           Test.Consensus.PeerSimulator.Resources
+import           Test.Consensus.PeerSimulator.Trace
+import qualified Test.Consensus.PointSchedule as PointSchedule
+import           Test.Consensus.PointSchedule (GenesisTest (GenesisTest),
+                     Peer (Peer), PeerId, PointSchedule (PointSchedule),
+                     TestFragH, Tick (Tick), pointSchedulePeers)
 import           Test.Ouroboros.Consensus.ChainGenerator.Params (Asc)
-import           Test.Ouroboros.Consensus.Genesis.Setup.GenChains (GenesisTest)
-import qualified Test.Ouroboros.Consensus.PeerSimulator.BlockFetch as PeerSimulator.BlockFetch
-import           Test.Ouroboros.Consensus.PeerSimulator.Config
-import           Test.Ouroboros.Consensus.PeerSimulator.Resources
-import           Test.Ouroboros.Consensus.PeerSimulator.Trace
-import qualified Test.Ouroboros.Consensus.PointSchedule as PointSchedule
-import           Test.Ouroboros.Consensus.PointSchedule
-                     (GenesisTest (GenesisTest), Peer (Peer), PeerId,
-                     PointSchedule (PointSchedule), TestFragH, Tick (Tick),
-                     pointSchedulePeers)
 import           Test.Util.ChainDB
 import           Test.Util.Orphans.IOLike ()
 import           Test.Util.TestBlock (Header (..), TestBlock, testInitExtLedger)
