+++ ouroboros-consensus-diffusion/test/consensus-test/Test/Consensus/PeerSimulator/Trace.hs
@@ -2,7 +2,7 @@
 {-# LANGUAGE NamedFieldPuns #-}
 
 -- | Helpers for tracing used by the peer simulator.
-module Test.Ouroboros.Consensus.PeerSimulator.Trace (
+module Test.Consensus.PeerSimulator.Trace (
     mkCdbTracer
   , mkChainSyncClientTracer
   , traceUnitWith
