+++ ouroboros-consensus/ouroboros-consensus.cabal
@@ -517,55 +517,23 @@ test-suite infra-test
   main-is:        Main.hs
   other-modules:
     Ouroboros.Consensus.Util.Tests
-    Ouroboros.Network.AnchoredFragment.Extras
-    Ouroboros.Network.Driver.Limits.Extras
-    Test.Ouroboros.Consensus.BlockTree
     Test.Ouroboros.Consensus.ChainGenerator.Tests
     Test.Ouroboros.Consensus.ChainGenerator.Tests.Adversarial
     Test.Ouroboros.Consensus.ChainGenerator.Tests.BitVector
     Test.Ouroboros.Consensus.ChainGenerator.Tests.Counting
     Test.Ouroboros.Consensus.ChainGenerator.Tests.Honest
-    Test.Ouroboros.Consensus.Genesis.Setup
-    Test.Ouroboros.Consensus.Genesis.Setup.Classifiers
-    Test.Ouroboros.Consensus.Genesis.Setup.GenChains
-    Test.Ouroboros.Consensus.Genesis.Tests
-    Test.Ouroboros.Consensus.Genesis.Tests.LongRangeAttack
-    Test.Ouroboros.Consensus.PeerSimulator.BlockFetch
-    Test.Ouroboros.Consensus.PeerSimulator.Config
-    Test.Ouroboros.Consensus.PeerSimulator.Handlers
-    Test.Ouroboros.Consensus.PeerSimulator.Resources
-    Test.Ouroboros.Consensus.PeerSimulator.Run
-    Test.Ouroboros.Consensus.PeerSimulator.ScheduledChainSyncServer
-    Test.Ouroboros.Consensus.PeerSimulator.Trace
-    Test.Ouroboros.Consensus.PointSchedule
     Test.Util.ChainUpdates.Tests
     Test.Util.Schedule.Tests
     Test.Util.Split.Tests
 
   build-depends:
     , base
-    , cardano-crypto-class
-    , cardano-slotting
-    , containers
-    , contra-tracer
-    , hashable
-    , io-classes
-    , io-sim
     , mtl
     , ouroboros-consensus:{ouroboros-consensus, unstable-consensus-testlib}
-    , ouroboros-network
-    , ouroboros-network-api
-    , ouroboros-network-framework
-    , ouroboros-network-protocols
     , QuickCheck
     , random
-    , si-timers
-    , strict-stm
     , tasty
     , tasty-quickcheck
-    , time
-    , typed-protocols
-    , typed-protocols-examples
     , unstable-consensus-testlib
     , vector
