+++ ouroboros-consensus/test/infra-test/Main.hs
@@ -14,7 +14,6 @@ module Main (main) where
 
 import qualified Ouroboros.Consensus.Util.Tests (tests)
 import qualified Test.Ouroboros.Consensus.ChainGenerator.Tests (tests)
-import qualified Test.Ouroboros.Consensus.Genesis.Tests (tests)
 import           Test.Tasty (TestTree, testGroup)
 import qualified Test.Util.ChainUpdates.Tests (tests)
 import qualified Test.Util.Schedule.Tests (tests)
@@ -30,7 +29,6 @@ tests =
   testGroup "test-infra"
   [ Ouroboros.Consensus.Util.Tests.tests
   , Test.Ouroboros.Consensus.ChainGenerator.Tests.tests
-  , Test.Ouroboros.Consensus.Genesis.Tests.tests
   , Test.Util.ChainUpdates.Tests.tests
   , Test.Util.Schedule.Tests.tests
   , Test.Util.Split.Tests.tests
